+++ func/checkIng.php
@@ -1,15 +1,15 @@
 <?php if (!defined('pvault_panel')){ die('Not Found');}?>
 <?php
-function checkIng($ingredient,$conn){
-	$chk = mysqli_query($conn, "SELECT name, IFRA, price, profile, cas FROM ingredients WHERE name = '$ingredient' OR chemical_name = '$ingredient'");	
+function checkIng($ingredient, $defCatClass, $conn){
+	$chk = mysqli_query($conn, "SELECT name, $defCatClass, price, profile, cas FROM ingredients WHERE name = '$ingredient' OR chemical_name = '$ingredient'");	
 	if(mysqli_num_rows($chk)){
 		while ($qValues=mysqli_fetch_array($chk)){
 			if($qValues['cas']){
 				$casQ = "OR cas LIKE '%$qValues[cas]%'";
 			}
-			$chkIFRA = mysqli_fetch_array(mysqli_query($conn, "SELECT name,cat4 FROM IFRALibrary WHERE  name = '$qValues[name]' OR synonyms LIKE '%$qValues[name]%' $casQ"));
+			$chkIFRA = mysqli_fetch_array(mysqli_query($conn, "SELECT name, $defCatClass FROM IFRALibrary WHERE  name = '$qValues[name]' OR synonyms LIKE '%$qValues[name]%' $casQ"));
 						
-			if(empty($chkIFRA['cat4']) && empty($qValues['IFRA'])){
+			if(empty($chkIFRA[$defCatClass]) && empty($qValues[$defCatClass])){
 				return '<a href="#" class="fas fa-exclamation" rel="tipsy" title="Missing usage data"></a>';
 			}elseif(empty($qValues['price'])){
 				return '<a href="#" class="fas fa-exclamation" rel="tipsy" title="Missing price data"></a>';